Fixed Install bug, Added p10k config

+++ install.sh
@@ -1,8 +1,15 @@
-
-echo "Wellcome to DOTFILES"
+echo "Welcome to DOTFILES"
 echo "Install:"
-echo "	- tmux"
-echo "	- zsh"
+echo "  - tmux"
+echo "  - zsh"
 
-rm $HOME/.zshrc
+if ! [ -d "$HOME/.zshrc" ] ; then
+    rm $HOME/.zshrc
+fi
 ln -s $HOME/.dotfiles/.zshrc $HOME/.zshrc
+
+if ! [ -d "$HOME/.p10k.zsh" ] ; then
+    rm $HOME/.p10k.zsh
+fi
+ln -s $HOME/.dotfiles/.p10k.zsh $HOME/.p10k.zsh
+
